+++ ANDOR-CS/Enums/EMGainEnum.cs
@@ -9,12 +9,12 @@ using SDK = ATMCD64CS.AndorSDK;
 namespace ANDOR_CS.Enums
 {
     [Flags]
-    public enum EMGain : int
+    public enum EMGain : uint
     {
         Unknown = 0,
-        Bits8 = (int)SDK.AC_EMGAIN_8BIT,
-        Bits12 = (int)SDK.AC_EMGAIN_12BIT,
-        LinearBits12 = (int)SDK.AC_EMGAIN_LINEAR12,
-        RealBits12 = (int)SDK.AC_EMGAIN_REAL12
+        Bits8 = SDK.AC_EMGAIN_8BIT,
+        Bits12 = SDK.AC_EMGAIN_12BIT,
+        LinearBits12 = SDK.AC_EMGAIN_LINEAR12,
+        RealBits12 = SDK.AC_EMGAIN_REAL12
     }
 }
